+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - "preload_models" option in config.json
 - Added batch for **Video Analyzer**
 - Lazy mode batch **Video Analyzer** (skiping per x frames)
+- Threshold slider for **Video Analyzer**
 ## 1.5.x
 - *New* system of loading models (35s preloading, but faster working)
 - Code optimization